split cli_test into smaller modules
